+++ index.php
@@ -7,7 +7,7 @@
     this file. It initializes the rest of the system.
 
     @package urlaube\urlaube
-    @version 0.1a0
+    @version 0.1a1
     @author  Yahe <hello@yahe.sh>
     @since   0.1a0
   */