+++ router.php
@@ -7,7 +7,7 @@
     Usage: `php -S localhost:8080 ./router.php`
 
     @package urlaube\urlaube
-    @version 0.1a0
+    @version 0.1a1
     @author  Yahe <hello@yahe.sh>
     @since   0.1a0
   */