+++ system/core/Theme.interface.php
@@ -7,7 +7,7 @@
     theme should implement to simply the implementation of further functions
 
     @package urlaube\urlaube
-    @version 0.1a0
+    @version 0.1a1
     @author  Yahe <hello@yahe.sh>
     @since   0.1a0
   */