+++ system/core/Translatable.class.php
@@ -8,7 +8,7 @@
     translation files, storing loaded translations and reading the translated strings.
 
     @package urlaube\urlaube
-    @version 0.1a0
+    @version 0.1a1
     @author  Yahe <hello@yahe.sh>
     @since   0.1a0
   */