+++ system/handlers/100_favicon_ico/handler.php
@@ -7,7 +7,7 @@
     contents for certain typically provided files.
 
     @package urlaube\urlaube
-    @version 0.1a0
+    @version 0.1a1
     @author  Yahe <hello@yahe.sh>
     @since   0.1a0
   */