+++ system/handlers/200_addslash/handler.php
@@ -7,7 +7,7 @@
     generic URL structure for all other handlers.
 
     @package urlaube\urlaube
-    @version 0.1a0
+    @version 0.1a1
     @author  Yahe <hello@yahe.sh>
     @since   0.1a0
   */