+++ system/handlers/300_archive/handler.php
@@ -7,7 +7,7 @@
     archive handler lists all pages that contain a certain date.
 
     @package urlaube\urlaube
-    @version 0.1a0
+    @version 0.1a1
     @author  Yahe <hello@yahe.sh>
     @since   0.1a0
   */