+++ system/handlers/310_category/handler.php
@@ -8,7 +8,7 @@
     identifier.
 
     @package urlaube\urlaube
-    @version 0.1a0
+    @version 0.1a1
     @author  Yahe <hello@yahe.sh>
     @since   0.1a0
   */