+++ system/handlers/370_search/handler.php
@@ -7,7 +7,7 @@
     search handler lists all pages that contain a certain search keyword.
 
     @package urlaube\urlaube
-    @version 0.1a0
+    @version 0.1a1
     @author  Yahe <hello@yahe.sh>
     @since   0.1a0
   */