+++ system/handlers/400_error/handler.php
@@ -7,7 +7,7 @@
     activated when no fitting handler is found.
 
     @package urlaube\urlaube
-    @version 0.1a0
+    @version 0.1a1
     @author  Yahe <hello@yahe.sh>
     @since   0.1a0
   */