+++ system/init.php
@@ -7,7 +7,7 @@
     kickstarts the execution of the CMS.
 
     @package urlaube\urlaube
-    @version 0.1a0
+    @version 0.1a1
     @author  Yahe <hello@yahe.sh>
     @since   0.1a0
   */
@@ -26,37 +26,37 @@
   require_once(SYSTEM_PATH."defaults.php");
   require_once(SYSTEM_PATH."recommends.php");
 
-  // require system core base classes
+  // require base classes
   require_once(SYSTEM_CORE_PATH."Base.class.php");
   require_once(SYSTEM_CORE_PATH."Content.class.php");
   require_once(SYSTEM_CORE_PATH."Debug.class.php");
 
-  // require system core main class
+  // require main class
   require_once(SYSTEM_CORE_PATH."Main.class.php");
 
-  // require system core classes
+  // require management classes
   require_once(SYSTEM_CORE_PATH."Handlers.class.php");
   require_once(SYSTEM_CORE_PATH."Plugins.class.php");
   require_once(SYSTEM_CORE_PATH."Themes.class.php");
   require_once(SYSTEM_CORE_PATH."Translations.class.php");
 
-  // require system core interfaces
+  // require interfaces
   require_once(SYSTEM_CORE_PATH."Handler.interface.php");
   require_once(SYSTEM_CORE_PATH."Plugin.interface.php");
   require_once(SYSTEM_CORE_PATH."Theme.interface.php");
   require_once(SYSTEM_CORE_PATH."Translation.interface.php");
 
-  // require system core default translation implementation
+  // require default translation implementation
   require_once(SYSTEM_CORE_PATH."Translatable.class.php");
 
-  // require system core config class
+  // require config class
   require_once(SYSTEM_CORE_PATH."Config.class.php");
 
-  // require the user configuration
+  // require user configuration
   if (is_file(USER_CONFIG_PATH.CONFIG_FILE)) {
     require_once(USER_CONFIG_PATH.CONFIG_FILE);
   }
 
-  // transfer the handling to the Main class
+  // transfer handling to the Main class
   print(Main::run());