+++ system/plugins/000_file/plugin.php
@@ -7,7 +7,7 @@
     CMS entries.
 
     @package urlaube\urlaube
-    @version 0.1a0
+    @version 0.1a1
     @author  Yahe <hello@yahe.sh>
     @since   0.1a0
   */