+++ system/recommends.php
@@ -7,7 +7,7 @@
     theme developers so that they can decide on often-used placeholders.
 
     @package urlaube\urlaube
-    @version 0.1a0
+    @version 0.1a1
     @author  Yahe <hello@yahe.sh>
     @since   0.1a0
   */